requirements.txt
Django>=4.0.0
django-cors-headers=3.12.0 
djangorestframework=3.13.1
